refactor: 폴더 및 컴포넌트명 변경, 주석작업
