+++ front-end/src/pages/garden/GardenPage.jsx
@@ -1,12 +1,21 @@
 import styled from "styled-components";
-import TestPage from "../component/TestPage";
-import GardenFooter from "../components/GardenFooter";
+
+import GardenFirst from "./GardenFirst";
+import GardenSecond from "./GardenSecond";
 
 export default function GardenPage() {
     return (
-        <div>
-            <TestPage />
-            <GardenFooter />
-        </div>
+        <Wrapper>
+            {/** TODO
+             * 현재 garden_id가 무엇인지에 따라 서로 다른 정원 데이터 불러오기
+             * id에 해당하는 건 First, 그 외는 Second
+             */}
+            <GardenFirst />
+            <GardenSecond />
+        </Wrapper>
     )
 }
+
+const Wrapper = styled.div`
+    width: 1280px;
+`
